Adicionado estruturas condicionais e repetição

+++ estruturas-condicionais/if-ternario.py
@@ -1,5 +1,5 @@
 saldo = 1000
-saque = int(input('Valor '))
+saque = float(input('Valor '))
 restante = saldo - saque
 
 status = 'Sucesso' if saldo >= saque else 'Falha'
